Modulated bass shift around C
Previously, changing the note parameter would simply add the difference
to the bass pitch. This lead to bass pitches that clashed a bit with the
range of the melody.

Instead, the pitch difference now wraps to a negative value, lowering
the bass pitch if the shift would push it too far above.